update validation for language APIs

+++ backend/controllers/languageControllers.js
@@ -2,9 +2,40 @@ import * as languageServices from '../services/languageServices';
 import { logger } from '../shared/utils/loggerUtilities';
 import { LanguageDTO } from '../dto/LanguageDTO';
 import HttpResponseResult from '../shared/models/HttpResponseResult';
+import { check, validationResult } from 'express-validator';
+
+export const validate = (method) => {
+    switch (method) {
+        case 'createLanguage': {
+            return [
+                check('user_id', 'userid is empty').not().isEmpty(),
+                check('language', 'language is empty').not().isEmpty(),
+                check('proficiency', 'proficiency is empty').not().isEmpty(),
+            ]
+        }
+        case 'deleteLanguage': {
+            return [
+                check('id', 'id is empty').not().isEmpty(),
+            ]
+        }
+        case 'updateLanguage': {
+            return [
+                check('id', 'id is empty').not().isEmpty(),
+                check('user_id', 'userid is empty').not().isEmpty(),
+                check('language', 'language is empty').not().isEmpty(),
+                check('proficiency', 'proficiency is empty').not().isEmpty(),
+            ]
+        }
+    }
+}
 
 export const add = async (req, res) => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(422).json({ errors: errors.array() });
+        }
+
         let data = req.body;
         let language = new LanguageDTO(data.user_id, data.language, data.proficiency);
         let createdLanguage = await languageServices.add(language);
@@ -43,6 +74,11 @@ export const detail = async (req, res) => {
 
 export const deleteById = async (req, res) => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(422).json({ errors: errors.array() });
+        }
+
         let language = await languageServices.deleteById(req.body.id);
 
         return await res.status(200).send(new HttpResponseResult(true, "", language));
@@ -59,6 +95,11 @@ export const deleteById = async (req, res) => {
 
 export const updateById = async (req, res) => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(422).json({ errors: errors.array() });
+        }
+
         let data = req.body;
         let language = new LanguageDTO(data.user_id, data.language, data.proficiency);
         let updatedLanguage = await languageServices.updateById(data.id, language);
